avoid importing numba unless used
